+++ backend/storage/migrations/2025-07-25.10.00-fix-settings.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `mythicaldash_settings` ADD UNIQUE(`name`);